persistent settings and landscape mode
landscape is not possible, looks fine in preview, only landscape left and portrait upright seems to work when running the app though.

Settings are saved vid UserDefaults through Appstorage. Ugly fix to set values if its the first ever launch of the app in the init method in the VM.